+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added a way to resize the vocabulary in the T5 module
+
 ### Fixed
 
 - Fixed the docstring information for the `FBetaMultiLabelMeasure` metric.